refactor: remove deprecated shim modules for unified_quarantine and filter_config
Remove deprecated backward-compatibility shims that are no longer needed:

- Remove src/bioetl/infrastructure/quarantine/unified_quarantine.py
  - Was a re-export shim for UnifiedQuarantine
  - Updated 4 files to import from canonical location

- Remove src/bioetl/domain/filter_config.py
  - Was a re-export shim for filtering package
  - Updated 15+ files to import from bioetl.domain.filtering

Changes:
- Update quarantine/__init__.py to export quote_literal helper
- Update all imports to use canonical package paths
- Update test files to use new import locations

Note: DeltaWriter deprecation shim (delta_writer.py) is kept as its
14-day deprecation period just started (RULES.md §6.2).

All 1620 unit tests pass. Lint and mypy pass.
