+++ src/bioetl/composition/factories/data_source_factory.py
@@ -24,7 +24,7 @@ from bioetl.composition.providers import (
 )
 
 if TYPE_CHECKING:
-    from bioetl.domain.filter_config import InputFilterConfig
+    from bioetl.domain.filtering import InputFilterConfig
     from bioetl.domain.ports import DataSourcePort, LoggerPort, MetricsPort
     from bioetl.infrastructure.adapters.http.client import UnifiedHTTPClient
     from bioetl.infrastructure.config import Settings
