--- src/bioetl/domain/filter_config.py
@@ -1,29 +1,0 @@
-"""Filter configuration for pipeline filtering.
-
-This module is a compatibility shim that re-exports from the filtering package.
-For new code, prefer importing directly from bioetl.domain.filtering.
-
-Defines the configuration for:
-- Input filtering: API requests based on input IDs from external sources (CSV files)
-- Gold filtering: Configurable column-based filters for Gold layer records
-"""
-
-from bioetl.domain.filtering import (
-    FilterLoadResult,
-    GoldColumnFilter,
-    GoldFilterConfig,
-    GoldListContainsFilter,
-    GoldListLengthFilter,
-    GoldRangeFilter,
-    InputFilterConfig,
-)
-
-__all__ = [
-    "FilterLoadResult",
-    "GoldColumnFilter",
-    "GoldFilterConfig",
-    "GoldListContainsFilter",
-    "GoldListLengthFilter",
-    "GoldRangeFilter",
-    "InputFilterConfig",
-]
