+++ src/bioetl/infrastructure/quarantine/__init__.py
@@ -5,6 +5,7 @@ Implements RULES.md §2.6 - Quarantine Policy.
 
 from __future__ import annotations
 
+from bioetl.infrastructure.quarantine.helpers import quote_literal
 from bioetl.infrastructure.quarantine.unified import UnifiedQuarantine
 
-__all__ = ["UnifiedQuarantine"]
+__all__ = ["UnifiedQuarantine", "quote_literal"]
