+++ tests/unit/domain/test_filter_config.py
@@ -4,7 +4,7 @@ from __future__ import annotations
 
 import pytest
 
-from bioetl.domain.filter_config import (
+from bioetl.domain.filtering import (
     GoldColumnFilter,
     GoldFilterConfig,
     InputFilterConfig,
